rs/telegram/requirements.txt
python-dotenv
python-telegram-bot
logdecorator